feat(sql): implement sink configs for s3, lancedb, filesystem, delta, and iceberg (#226)
Add typed sink configuration building and validation for S3 and LanceDB,
and complete passthrough sink wiring for filesystem, delta, and iceberg
connectors. This change focuses on sink-side option parsing, runtime
property propagation, and connector-specific format checks.

+++ src/runtime/streaming/api/operator.rs
@@ -66,7 +66,7 @@ pub trait Operator: Send + 'static {
     /// `commit_transaction` on the producer stashed during [`Self::snapshot_state`].
     async fn commit_checkpoint(
         &mut self,
-        epoch: u32,
+        epoch: u64,
         _ctx: &mut TaskContext,
     ) -> anyhow::Result<()> {
         let _ = epoch;
@@ -76,7 +76,7 @@ pub trait Operator: Send + 'static {
     /// Global checkpoint **rollback** when phase 2 must not commit (e.g. catalog persist failed).
     ///
     /// Default is no-op. Transactional Kafka sink overrides with `abort_transaction` on the stashed producer.
-    async fn abort_checkpoint(&mut self, epoch: u32, _ctx: &mut TaskContext) -> anyhow::Result<()> {
+    async fn abort_checkpoint(&mut self, epoch: u64, _ctx: &mut TaskContext) -> anyhow::Result<()> {
         let _ = epoch;
         Ok(())
     }
